recursively check evaluations
When we do a job update, this is usually a rolling update. Therefore
Nomad will evaluate n allocations (depending on how the update is
configured) at a time. To monitor all rolling allocations we need to
inspect the `NextEval` identifier. It this is set, we need to follow
the chain.

As an improvement we should monitor the allocation in the background
as soon as the evaluation finished. This allows us to verify allocations
aren't failing as soon as possible. As it currently stands we could
monitor all evaluations up-front, just to find out afterwards that all
allocations have failed.