fix race condition in engine test
Race detector was pancking due to the fact we were attempting to
read the producer message whilst it was being modified. Flipped the
code around to cancel the context after a short period of time
(enough to allow the consumer to consume messages), and then
validating the correct response is returned once the consumer has
exited.