limit the number of messages we concurrently handle
This is more of a safeguard to prevent a run-away train e.g if our
CI really poops the bed (or something else, like writing a million
messages to the queue with a misconfigured script). We should be
able to run this many handlers concurrently, but we may not provide
the grace time required for them all to complete before the process
is sigkilled.