vendor deps
I attempted to use `dep` for this, but omg, the issues with resolving
dependencies was endless (was trying to wrangle specific versions of
Docker that Nomad depends on, omg). Might be better to try `dep` on
something which doesn't depend on alot of libs (vault will be required
also soon, which just adds more to the pile).