set credential file location to a silly value
Nathan spotted this issue which causes the test to fail if the user
has an AWS credential file configured. Setting it to a stupid path
fixes the test for me when I have a credentials file set-up. Hopefully
nobody will be storying a credentials file in `/i/hope/this/path/does/not/exist`.