use a http client with a timeout
Ten seconds seems high, but this is the timeout for the entire request
to complete (even consuming the request body). There is a way that we
could tweak the timeout values more specifically (time until first byte
recived etc), but I think this crude timeout will suffice for the time
being.