remove concatenated errors
This was a side-effect of writing an informative message back to the queue.

Removing the concatenation in error strings lead to having to restructure
the queue response slightly so that errors are also structued within the
queue response messages. We now write the error string as well as a JSON
representation of the error struct itself.