Add ability to explicitly skip TLS verification
As the deployer is now running inside a docker container it will not be
able to access Nomad on `localhost`.  Therefore, only disabling TLS
verification on connections that include localhost is not enough.  This
change adds a configuration to enable the user to explicitly skip the
TLS verification if desired.