Merge sqs lib in. Place struct in channel. Unmarshal into struct.
Merge in the version of ssqs lib that is referenced in the mod/sum files, which happens not to be master - thus fixing confusions.
(the ssqs lib is only used by deployer - so should be removed).
Send struct in channel instead of pointer.
Unpack into struct and not into pointer.
Rate limit polling of SQS receive messages instead os spinning.