Attempt to fix static port parsing in nomad plans
When specifying static ports in nomad plans this configuration is being
lost. When planning these deployments directly via the nomad ui they
work as expected and the static port is configured for the task group.
When planning via the deployer, these apps end up being deployed with
dynamic ports.

Attempting to rollback to the latest 0.12.x release to match the version
of nomad we are running in the hopes this may address the issue.

This change will be rolled back if testing in sandbox does not show the
issues as fixed.